# swagger_json.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/audit/logs": {
      "get": {
        "operationId": "AuditLogController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "description": "操作类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceType",
            "required": false,
            "in": "query",
            "description": "资源类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "required": false,
            "in": "query",
            "description": "资源 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "开始日期",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "结束日期",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "success",
            "required": false,
            "in": "query",
            "description": "是否成功",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "example": 20,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取审计日志"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询审计日志",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/logs/{id}": {
      "get": {
        "operationId": "AuditLogController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取审计日志详情"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取审计日志详情",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/statistics": {
      "get": {
        "operationId": "AuditLogController_getStatistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "开始日期",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "结束日期",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取审计统计信息"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取审计统计信息",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/audit/cleanup": {
      "post": {
        "operationId": "AuditLogController_cleanupOldLogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功清理旧审计日志"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "清理旧审计日志",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "operationId": "AdminController_getAdminStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取管理员统计信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取管理员统计信息",
        "tags": [
          "管理员"
        ]
      }
    },
    "/api/admin/permissions/cache": {
      "get": {
        "operationId": "AdminController_getCacheStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取权限缓存统计成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheStatsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取权限缓存统计",
        "tags": [
          "管理员"
        ]
      }
    },
    "/api/admin/permissions/cache/cleanup": {
      "post": {
        "operationId": "AdminController_cleanupCache",
        "parameters": [],
        "responses": {
          "200": {
            "description": "缓存清理完成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheCleanupResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "清理权限缓存",
        "tags": [
          "管理员"
        ]
      }
    },
    "/api/admin/permissions/cache/user/{userId}": {
      "delete": {
        "operationId": "AdminController_clearUserCache",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户权限缓存已清除",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCacheClearResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "清除用户权限缓存",
        "tags": [
          "管理员"
        ]
      }
    },
    "/api/admin/permissions/user/{userId}": {
      "get": {
        "operationId": "AdminController_getUserPermissions",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取用户权限信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPermissionsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户权限信息",
        "tags": [
          "管理员"
        ]
      }
    },
...
  },
  "info": {
    "title": "CloudCAD API",
    "description": "图纸管理平台API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminStatsResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          },
          "timestamp": {
            "type": "string",
            "description": "时间戳"
          }
        },
        "required": [
          "message",
          "timestamp"
        ]
      },
      "CacheStatsDto": {
        "type": "object",
        "properties": {
          "size": {
            "type": "number",
            "description": "缓存条目数"
          },
          "hits": {
            "type": "number",
            "description": "命中次数"
          },
          "misses": {
            "type": "number",
            "description": "未命中次数"
          },
          "hitRate": {
            "type": "number",
            "description": "命中率"
          }
        },
        "required": [
          "size",
          "hits",
          "misses",
          "hitRate"
        ]
      },
      "CacheStatsResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          },
          "data": {
            "description": "缓存统计数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/CacheStatsDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
      "CacheCleanupResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserCacheClearResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserPermissionInfoDto": {
        "type": "object",
        "properties": {
          "userRole": {
            "type": "string",
            "description": "用户角色"
          },
          "permissions": {
            "description": "权限列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userRole",
          "permissions"
        ]
      },
      "UserPermissionsResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "提示消息"
          },
          "data": {
            "description": "用户权限信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserPermissionInfoDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
    }
  }
}